requirements.txt
Flask>=2.0,<3
gunicorn>=20
playwright>=1.40.0
notion-client>=0.1.3
google-api-python-client>=2.70.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
python-json-logger>=2.0.7
pydantic>=1.10.0
requests>=2.28.1
pytest>=7.0.0  # For running unit tests